Removed LdapUserAuthorityPopulator and instead use CustomUserAuthoritiesMapper.
